+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### [Latest]
 
+- Enable Writing of Multiple Output Files in Merging [#97](https://github.com/umami-hep/umami-preprocessing/pull/97)
 - Rewrite Plotting Code to Enhance Speed [#96](https://github.com/umami-hep/umami-preprocessing/pull/96)
 - Add example config for open dataset [#92](https://github.com/umami-hep/umami-preprocessing/pull/92)
 - Adding GN3V01 Configs & renaming old Folder [#95](https://github.com/umami-hep/umami-preprocessing/pull/95)